+++ packages/common/src/hooks/useBreadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export * from './useBreadcrumbs';
